+++ pytest_factoryboy/plugin.py
@@ -9,28 +9,30 @@ class Request(object):
 
     def __init__(self):
         self.deferred = []
-        self.is_evaluated = False
+        self.is_finalized = False
 
     def defer(self, function):
         """Defer post-generation declaration execution until the end of the test setup.
 
         :param function: Function to be deferred.
-        :note: Once already evaluated all following defer calls will execute the function directly.
+        :note: Once already finalized all following defer calls will execute the function directly.
         """
-        if self.is_evaluated:
+        if self.is_finalized:
             function()
         else:
             self.deferred.append(function)
 
-    def evaluate(self):
+    def evaluate(self, names=None):
         """Finalize, run deferred post-generation actions, etc."""
-        while True:
-            try:
-                self.deferred.pop(0)()
-            except IndexError:
-                return
-            finally:
-                self.is_evaluated = True
+        if names:
+            functions = dict((function.__name__, function) for function in self.deferred if function.__name__ in names)
+            deferred = [functions[name] for name in names if functions.get(name)]
+        else:
+            deferred = list(self.deferred)
+
+        for function in deferred:
+            function()
+            self.deferred.remove(function)
 
 
 @pytest.fixture
@@ -47,7 +49,9 @@ def pytest_runtest_call(item):
     except AttributeError:
         # pytest-pep8 plugin passes Pep8Item here during tests.
         return
-    request.getfuncargvalue("factoryboy_request").evaluate()
+    factoryboy_request = request.getfuncargvalue("factoryboy_request")
+    factoryboy_request.evaluate()
+    factoryboy_request.is_finalized = True
     request.config.hook.pytest_factoryboy_done(request=request)